refactor(tui): split the 3529-LOC harness.rs test module into concern files
Issue #482: the single `tui_renderer/harness.rs` test file (3529 LOC, 104 tests)
became a directory module `harness/` of 10 concern-area files (floors,
theme_lighting, coffee_pets, mascot, hit_test, overlays, render_text,
edge_cases, dashboard + the shared-helper mod.rs). The `#[cfg(test)] mod
harness;` declaration resolves to harness/mod.rs unchanged; NO production code
touched.

Move-only: all 104 test bodies are byte-identical (verified by extracting all
127 fns old vs new and diffing each block — the only non-verbatim change is the
`slot` HELPER's signature, rustfmt-wrapped by the `pub(super)` prefix). Shared
helpers live in mod.rs as `pub(super)`; single-group helpers localized private.
Docs-currency: the CLAUDE.md refs to `harness.rs` repointed to the `harness`
module.

104 harness tests pass; clippy -D warnings clean; net +26 structural lines.
